slack-notifier/requirements.txt
functions-framework==3.8.1
google-cloud-secret-manager==2.20.2
